+++ src/main/java/com/solace/samples/jcsmp/samplers/AdPubAck.java
@@ -37,10 +37,10 @@
 package com.solace.samples.jcsmp.samplers;
 
 import java.util.LinkedList;
-import com.solace.samples.samplers.common.ArgParser;
-import com.solace.samples.samplers.common.SampleApp;
-import com.solace.samples.samplers.common.SampleUtils;
-import com.solace.samples.samplers.common.SessionConfiguration;
+import com.solace.samples.samplers.jcsmp.common.ArgParser;
+import com.solace.samples.samplers.jcsmp.common.SampleApp;
+import com.solace.samples.samplers.jcsmp.common.SampleUtils;
+import com.solace.samples.samplers.jcsmp.common.SessionConfiguration;
 import com.solacesystems.jcsmp.BytesXMLMessage;
 import com.solacesystems.jcsmp.DeliveryMode;
 import com.solacesystems.jcsmp.JCSMPException;
