+++ src/main/java/com/solace/samples/jcsmp/samplers/TopicToQueueMapping.java
@@ -8,10 +8,10 @@
 
 package com.solace.samples.jcsmp.samplers;
 
-import com.solace.samples.samplers.common.ArgParser;
-import com.solace.samples.samplers.common.SampleApp;
-import com.solace.samples.samplers.common.SampleUtils;
-import com.solace.samples.samplers.common.SessionConfiguration;
+import com.solace.samples.samplers.jcsmp.common.ArgParser;
+import com.solace.samples.samplers.jcsmp.common.SampleApp;
+import com.solace.samples.samplers.jcsmp.common.SampleUtils;
+import com.solace.samples.samplers.jcsmp.common.SessionConfiguration;
 import com.solacesystems.jcsmp.BytesXMLMessage;
 import com.solacesystems.jcsmp.CapabilityType;
 import com.solacesystems.jcsmp.Consumer;
