+++ src/main/java/com/solace/samples/samplers/jcsmp/common/SampleApp.java
@@ -4,7 +4,7 @@
  * Copyright 2009-2018 Solace Corporation. All rights reserved.
  */
 
-package com.solace.samples.samplers.common;
+package com.solace.samples.samplers.jcsmp.common;
 
 import java.util.ArrayList;
 import java.util.Arrays;
